fourth push

+++ static/js/slideshow.js
@@ -1,7 +1,7 @@
 const slideshow = document.querySelector('.gallery div.slide_container');
 const slides = document.querySelectorAll(".slides");
 
-const slideWidth = slides[0].offsetWidth + parseFloat(getComputedStyle(slides[0]).marginLeft);
+const slideWidth = slides[0].offsetWidth + parseInt(getComputedStyle(slides[0]).marginLeft);
 const slidesToScroll = 1;
 const scrollDistance = slidesToScroll * slideWidth;
 document.querySelector('.slideshow button.left').addEventListener('click', async(e)=>{
